+++ pytorch_lightning/callbacks/gradient_accumulation_scheduler.py
@@ -23,6 +23,7 @@ Trainer also calls ``optimizer.step()`` for the last indivisible step number.
 from typing import Dict
 
 from pytorch_lightning.callbacks.base import Callback
+from pytorch_lightning.utilities.exceptions import MisconfigurationException
 
 
 class GradientAccumulationScheduler(Callback):
@@ -31,6 +32,14 @@ class GradientAccumulationScheduler(Callback):
 
     Args:
         scheduling: scheduling in format {epoch: accumulation_factor}
+
+    Note:
+        The argument scheduling is a dictionary. Each key represent an epoch and
+        its associated accumulation factor value.
+        Warning: Epoch are zero-indexed c.f it means if you want to change
+        the accumulation factor after 4 epochs, set ``Trainer(accumulate_grad_batches={4: factor})``
+        or ``GradientAccumulationScheduler(scheduling={4: factor})``.
+        For more info check the example below.
 
     Raises:
         TypeError:
@@ -44,12 +53,13 @@ class GradientAccumulationScheduler(Callback):
         >>> from pytorch_lightning import Trainer
         >>> from pytorch_lightning.callbacks import GradientAccumulationScheduler
 
-        # at epoch 5 start accumulating every 2 batches
-        >>> accumulator = GradientAccumulationScheduler(scheduling={5: 2})
+        # from epoch 5, it starts accumulating every 2 batches. Here we have 4 instead of 5
+        because epoch (key) should be zero-indexed.
+        >>> accumulator = GradientAccumulationScheduler(scheduling={4: 2})
         >>> trainer = Trainer(callbacks=[accumulator])
 
         # alternatively, pass the scheduling dict directly to the Trainer
-        >>> trainer = Trainer(accumulate_grad_batches={5: 2})
+        >>> trainer = Trainer(accumulate_grad_batches={4: 2})
     """
 
     def __init__(self, scheduling: Dict[int, int]):
@@ -58,9 +68,15 @@ class GradientAccumulationScheduler(Callback):
         if not scheduling:  # empty dict error
             raise TypeError("Empty dict cannot be interpreted correct")
 
-        for key in scheduling:
-            if not isinstance(key, int) or not isinstance(scheduling[key], int):
-                raise TypeError("All epoches and accumulation factor must be integers")
+        if any(not isinstance(key, int) or key < 0 for key in scheduling):
+            raise MisconfigurationException(
+                f"Epoch should be an int greater than or equal to 0. Got {list(scheduling.keys())}."
+            )
+
+        if any(not isinstance(value, int) or value < 1 for value in scheduling.values()):
+            raise MisconfigurationException(
+                f"Accumulation factor should be an int greater than 0. Got {list(scheduling.values())}."
+            )
 
         minimal_epoch = min(scheduling.keys())
         if minimal_epoch < 0: