+++ tests/loggers/test_base.py
@@ -318,9 +318,9 @@ def test_log_hyperparams_key_collision(log_hyperparams_mock, tmpdir):
         limit_train_batches=0.1,
         limit_val_batches=0.1,
         num_sanity_val_steps=0,
-        checkpoint_callback=False,
+        enable_checkpointing=False,
         enable_progress_bar=False,
-        weights_summary=None,
+        enable_model_summary=False,
     )
     # there should be no exceptions raised for the same key/value pair in the hparams of both
     # the lightning module and data module
@@ -342,9 +342,9 @@ def test_log_hyperparams_key_collision(log_hyperparams_mock, tmpdir):
         limit_train_batches=0.1,
         limit_val_batches=0.1,
         num_sanity_val_steps=0,
-        checkpoint_callback=False,
+        enable_checkpointing=False,
         enable_progress_bar=False,
-        weights_summary=None,
+        enable_model_summary=False,
     )
     with pytest.raises(MisconfigurationException, match="Error while merging hparams"):
         trainer.fit(model, dm)
@@ -359,9 +359,9 @@ def test_log_hyperparams_key_collision(log_hyperparams_mock, tmpdir):
         limit_train_batches=0.1,
         limit_val_batches=0.1,
         num_sanity_val_steps=0,
-        checkpoint_callback=False,
+        enable_checkpointing=False,
         enable_progress_bar=False,
-        weights_summary=None,
+        enable_model_summary=False,
     )
     with pytest.raises(MisconfigurationException, match="Error while merging hparams"):
         trainer.fit(model, dm)